chore: make shareable types package
