+++ apps/api/package.json
@@ -31,6 +31,7 @@
     "@prisma/client": "^4.9.0",
     "@sentry/integrations": "^7.34.0",
     "@sentry/node": "^7.34.0",
+    "@supaglue/types": "workspace:*",
     "@swc/core": "^1.3.27",
     "@temporalio/activity": "^1.5.2",
     "@temporalio/client": "^1.5.2",
