+++ apps/api/syncs/entities/sync.ts
@@ -1,173 +1,4 @@
-// TODO: Should be shared. Consider putting all entities in same directory.
-type SalesforceObject =
-  | 'Account'
-  | 'Activity'
-  | 'AlternativePaymentMethod'
-  | 'ApiAnomalyEventStore'
-  | 'AppointmentInvitation'
-  | 'AppointmentInvitee'
-  | 'AppointmentTopicTimeSlot'
-  | 'Asset'
-  | 'AssetAction'
-  | 'AssetActionSource'
-  | 'AssetRelationship'
-  | 'AssetStatePeriod'
-  | 'AssignedResource'
-  | 'AssociatedLocation'
-  | 'AuthorizationForm'
-  | 'AuthorizationFormConsent'
-  | 'AuthorizationFormDataUse'
-  | 'AuthorizationFormText'
-  | 'BusinessBrand'
-  | 'Campaign'
-  | 'CampaignMember'
-  | 'CardPaymentMethod'
-  | 'WebCart'
-  | 'WebCartAdjustmentBasis'
-  | 'WebCartAdjustmentGroup'
-  | 'CartDeliveryGroup'
-  | 'CartItem'
-  | 'CartItemPriceAdjustment'
-  | 'CartTax'
-  | 'Case'
-  | 'CaseRelatedIssue'
-  | 'ChangeRequest'
-  | 'ChangeRequestRelatedIssue'
-  | 'ChangeRequestRelatedItem'
-  | 'CommSubscription'
-  | 'CommSubscriptionChannelType'
-  | 'CommSubscriptionConsent'
-  | 'CommSubscriptionTiming'
-  | 'ConsumptionRate'
-  | 'ConsumptionSchedule'
-  | 'Contact'
-  | 'ContactPointAddress'
-  | 'ContactPointConsent'
-  | 'ContactPointEmail'
-  | 'ContactPointPhone'
-  | 'ContactPointTypeConsent'
-  | 'ContactRequest'
-  | 'ContentVersion'
-  | 'Contract'
-  | 'ContractLineItem'
-  | 'Coupon'
-  | 'CredentialStuffingEventStore'
-  | 'CreditMemo'
-  | 'CreditMemoInvApplication'
-  | 'CreditMemoLine'
-  | 'Customer'
-  | 'DandBCompany'
-  | 'DataUseLegalBasis'
-  | 'DataUsePurpose'
-  | 'DigitalWallet'
-  | 'DuplicateRecordItem'
-  | 'DuplicateRecordSet'
-  | 'EmailMessage'
-  | 'EngagementChannelType'
-  | 'EngagementChannelWorkType'
-  | 'Entitlement'
-  | 'EntitlementContact'
-  | 'Event'
-  | 'FinanceBalanceSnapshot'
-  | 'FinanceTransaction'
-  | 'Image'
-  | 'Incident'
-  | 'IncidentRelatedItem'
-  | 'Individual'
-  | 'Invoice'
-  | 'InvoiceLine'
-  | 'Lead'
-  | 'LegalEntity'
-  | 'ListEmail'
-  | 'Location'
-  | 'LocationGroup'
-  | 'LocationGroupAssignment'
-  | 'Macro'
-  | 'MessagingSession'
-  | 'MessagingEndUser'
-  | 'EntityMilestone'
-  | 'OperatingHours'
-  | 'OperatingHoursHoliday'
-  | 'Opportunity'
-  | 'OpportunityContactRole'
-  | 'OpportunityLineItem'
-  | 'Order'
-  | 'OrderItem'
-  | 'PartyConsent'
-  | 'Payment'
-  | 'PaymentAuthorization'
-  | 'PaymentAuthAdjustment'
-  | 'PaymentGateway'
-  | 'PaymentGroup'
-  | 'PaymentLineInvoice'
-  | 'Pricebook2'
-  | 'PricebookEntry'
-  | 'Problem'
-  | 'ProblemRelatedItem'
-  | 'ProcessException'
-  | 'Product2'
-  | 'ProductAttribute'
-  | 'ProductAttributeSetProduct'
-  | 'ProductConsumptionSchedule'
-  | 'Promotion'
-  | 'PromotionMarketSegment'
-  | 'PromotionQualifier'
-  | 'PromotionSegment'
-  | 'PromotionSegmentSalesStore'
-  | 'PromotionTarget'
-  | 'LocationWaitlist'
-  | 'LocWaitlistMsgTemplate'
-  | 'LocationWaitlistedParty'
-  | 'QuickText'
-  | 'Recommendation'
-  | 'Refund'
-  | 'RefundLinePayment'
-  | 'ProblemIncident'
-  | 'ReportAnomalyEventStore'
-  | 'ResourceAbsence'
-  | 'ResourcePreference'
-  | 'ReturnOrder'
-  | 'ReturnOrderItemAdjustment'
-  | 'ReturnOrderItemTax'
-  | 'ReturnOrderLineItem'
-  | 'Scorecard'
-  | 'ScorecardAssociation'
-  | 'ScorecardMetric'
-  | 'Seller'
-  | 'ServiceAppointment'
-  | 'ServiceAppointmentGroup'
-  | 'ServiceContract'
-  | 'ServiceResource'
-  | 'ServiceResourceSkill'
-  | 'ServiceTerritory'
-  | 'ServiceTerritoryMember'
-  | 'ServiceTerritoryRelationship'
-  | 'ServiceTerritoryWorkType'
-  | 'SessionHijackingEventStore'
-  | 'Shift'
-  | 'ShiftEngagementChannel'
-  | 'ShiftWorkTopic'
-  | 'SkillRequirement'
-  | 'SocialPersona'
-  | 'WebStore'
-  | 'Task'
-  | 'TimeSlot'
-  | 'User'
-  | 'UserProvisioningRequest'
-  | 'VoiceCall'
-  | 'WorkOrder'
-  | 'WorkOrderLineItem'
-  | 'WorkPlan'
-  | 'WorkPlanTemplate'
-  | 'WorkPlanTemplateEntry'
-  | 'WorkProcedure'
-  | 'WorkProcedureStep'
-  | 'WorkStep'
-  | 'WorkStepTemplate'
-  | 'WorkType'
-  | 'WorkTypeExtension'
-  | 'WorkTypeGroup'
-  | 'WorkTypeGroupMember';
+import { SalesforceObject } from '@supaglue/types';
 
 type SalesforceCustomerIntegrationParams = {
   // If the developer allows the salesforce object to be selectable by the customer,
