+++ apps/cli/package.json
@@ -10,6 +10,7 @@
     "@sentry/integrations": "^7.34.0",
     "@sentry/node": "^7.34.0",
     "@supaglue/sdk": "workspace:*",
+    "@supaglue/types": "workspace:*",
     "axios": "^1.2.2",
     "chalk": "^4.1.2",
     "clui": "^0.3.6",
