--- packages/sdk/src/base.ts
@@ -1,14 +1,0 @@
-import { FieldMapping } from './field_mapping';
-
-export type BaseSyncConfig = {
-  name: string; // unique (e.g. ContactSync, LeadSync, AccountSync)
-
-  // some valid cron string
-  // TODO: we'll want to allow triggered sync runs down the line
-  cronExpression: string;
-
-  // TODO: support incremental
-  strategy: 'full_refresh';
-
-  defaultFieldMapping?: FieldMapping[];
-};
