--- packages/sdk/src/customer/base/base.ts
@@ -1,1 +1,0 @@
-export type BaseCustomerIntegration = object;
