--- packages/sdk/src/customer/base/index.ts
@@ -1,1 +1,0 @@
-export * from './salesforce';
