--- packages/sdk/src/customer/base/salesforce.ts
@@ -1,7 +1,0 @@
-import { SalesforceObjectConfig } from '../common/salesforce';
-import { BaseCustomerIntegration } from './base';
-
-export type SalesforceCustomerIntegration = BaseCustomerIntegration & {
-  type: 'salesforce';
-  objectConfig: SalesforceObjectConfig;
-};
