+++ packages/sdk/src/customer/destinations/salesforce.ts
@@ -1,8 +1,4 @@
-import { SalesforceCustomerIntegration } from '../base';
-
-export type SalesforceDestination = SalesforceCustomerIntegration & {
-  upsertKey: string; // ext_id
-};
+import { SalesforceDestination } from '@supaglue/types';
 
 export function salesforce(params: Omit<SalesforceDestination, 'type'>): SalesforceDestination {
   return {
