+++ packages/sdk/src/customer/sources/salesforce.ts
@@ -1,6 +1,4 @@
-import { SalesforceCustomerIntegration } from '../base';
-
-export type SalesforceSource = SalesforceCustomerIntegration;
+import { SalesforceSource } from '@supaglue/types';
 
 export function salesforce(params: Omit<SalesforceSource, 'type'>): SalesforceSource {
   return {
