+++ packages/sdk/src/field_mapping.ts
@@ -1,7 +1,4 @@
-export type FieldMapping = {
-  name: string;
-  field: string;
-};
+import { FieldMapping } from '@supaglue/types';
 
 export function defaultFieldMapping(fieldMapping: FieldMapping[]): FieldMapping[] {
   return fieldMapping;
