--- packages/sdk/src/internal/base/index.ts
@@ -1,2 +1,0 @@
-export * from './postgres';
-export * from './webhook';
