--- packages/sdk/src/internal/base/postgres.ts
@@ -1,11 +1,0 @@
-import { PostgresCredentials } from '../common/postgres';
-import { BaseInternalIntegration } from './base';
-
-export type PostgresInternalIntegration = BaseInternalIntegration & {
-  type: 'postgres';
-  config: {
-    credentials: PostgresCredentials;
-    table: string;
-    customerIdColumn: string;
-  };
-};
