+++ packages/types/tsconfig.json
@@ -1,0 +1,9 @@
+{
+  "extends": "@tsconfig/node18/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "declaration": true
+  },
+  "include": ["**/*.ts"],
+  "exclude": ["node_modules"]
+}
